+++ packages/client-next/src/__tests__/utils.test.ts
@@ -186,4 +186,25 @@ describe('setAuthParams', () => {
     expect(headers.get('baz')).toBeNull();
     expect(query.baz).toBe('Bearer foo');
   });
+
+  it('sets an API key in a cookie', async () => {
+    const auth = vi.fn().mockReturnValue('foo');
+    const headers = new Headers();
+    const query: Record<any, unknown> = {};
+    await setAuthParams({
+      auth,
+      headers,
+      query,
+      security: [
+        {
+          in: 'cookie',
+          name: 'baz',
+          type: 'apiKey',
+        },
+      ],
+    });
+    expect(auth).toHaveBeenCalled();
+    expect(headers.get('Cookie')).toBe('baz=foo');
+    expect(query).toEqual({});
+  });
 });