+++ packages/client-core/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hey-api/client-core
 
+## 0.1.4
+
+### Patch Changes
+
+- [#1850](https://github.com/hey-api/openapi-ts/pull/1850) [`fe43b88`](https://github.com/hey-api/openapi-ts/commit/fe43b889c20a2001f56e259f93f64851a1caa1d1) Thanks [@kelnos](https://github.com/kelnos)! - feat: add support for cookies auth
+
 ## 0.1.3
 
 ### Patch Changes